Add new RadiusClient implementation (with Mutex wrapper for socket_poll)
