+++ tests/test_mutex_client.rs
@@ -1,13 +1,13 @@
-use radius_rust::client::client::Client;
+use radius_rust::client::mutex_client::Client;
 use radius_rust::protocol::dictionary::Dictionary;
 use radius_rust::tools::{ integer_to_bytes, ipv4_string_to_bytes};
 
 
-// Test normal RadiusClient
+// Test Mutex RadiusClient
 #[test]
-fn test_client_auth_request() {
+fn test_mutex_client_auth_request() {
     let dictionary = Dictionary::from_file("./dict_examples/integration_dict").unwrap();
-    let mut client = Client::initialise_client(1812, 1813, 3799, dictionary, String::from("127.0.0.1"), String::from("secret"), 1, 2).unwrap();
+    let client     = Client::initialise_client(1812, 1813, 3799, dictionary, String::from("127.0.0.1"), String::from("secret"), 1, 2).unwrap();
 
     let nas_ip_addr_bytes    = ipv4_string_to_bytes("192.168.1.10").unwrap();
     let framed_ip_addr_bytes = ipv4_string_to_bytes("10.0.0.100").unwrap();
@@ -38,9 +38,9 @@ fn test_client_auth_request() {
 }
 
 #[test]
-fn test_client_acct_request() {
+fn test_mutex_client_acct_request() {
     let dictionary = Dictionary::from_file("./dict_examples/integration_dict").unwrap();
-    let mut client = Client::initialise_client(1812, 1813, 3799, dictionary, String::from("127.0.0.1"), String::from("secret"), 1, 2).unwrap();
+    let client     = Client::initialise_client(1812, 1813, 3799, dictionary, String::from("127.0.0.1"), String::from("secret"), 1, 2).unwrap();
 
     let nas_ip_addr_bytes    = ipv4_string_to_bytes("192.168.1.10").unwrap();
     let framed_ip_addr_bytes = ipv4_string_to_bytes("10.0.0.100").unwrap();
@@ -68,9 +68,9 @@ fn test_client_acct_request() {
 }
 
 #[test]
-fn test_client_coa_request() {
+fn test_mutex_client_coa_request() {
     let dictionary = Dictionary::from_file("./dict_examples/integration_dict").unwrap();
-    let mut client = Client::initialise_client(1812, 1813, 3799, dictionary, String::from("127.0.0.1"), String::from("secret"), 1, 2).unwrap();
+    let client     = Client::initialise_client(1812, 1813, 3799, dictionary, String::from("127.0.0.1"), String::from("secret"), 1, 2).unwrap();
 
     let attributes = vec![
         client.create_attribute_by_name("User-Name",          String::from("testing").into_bytes()).unwrap(),
